completed unit 7a lab

+++ Unity Projects/IntroScriptableObjectsProject/Assets/Scripts/IntData.cs
@@ -3,14 +3,14 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "New Int Data", menuName = "Int Data")]
 public class IntData : ScriptableObject
 {
-    public float value;
+    public int value;
 
-    public void SetValue(float num)
+    public void SetValue(int num)
     {
         value = num;
     }
 
-    public void UpdateValue(float num)
+    public void UpdateValue(int num)
     {
         value += num;
     }
